Unify web_endpoint, asgi_app, wsgi_app
