+++ client_test/decorator_test.py
@@ -1,7 +1,7 @@
 # Copyright Modal Labs 2023
 import pytest
 
-from modal import Stub, method, web_endpoint
+from modal import Stub, method, web_endpoint, asgi_app, wsgi_app
 from modal.exception import InvalidError
 
 
@@ -55,15 +55,26 @@ def test_method_forgot_parentheses():
     assert "method()" in str(excinfo.value)
 
 
-def test_web_endpoint_forgot_parentheses():
+def test_invalid_web_decorator_usage():
     stub = Stub()
 
-    with pytest.raises(InvalidError) as excinfo:
+    with pytest.raises(InvalidError, match="web_endpoint()"):
 
-        @stub.cls()
-        class XYZ:
-            @web_endpoint  # type: ignore
-            def f(self):
-                pass
+        @stub.function()  # type: ignore
+        @web_endpoint  # type: ignore
+        def my_handle():
+            pass
 
-    assert "web_endpoint()" in str(excinfo.value)
+    with pytest.raises(InvalidError, match="asgi_app()"):
+
+        @stub.function()  # type: ignore
+        @asgi_app  # type: ignore
+        def my_handle_asgi():
+            pass
+
+    with pytest.raises(InvalidError, match="wsgi_app()"):
+
+        @stub.function()  # type: ignore
+        @wsgi_app  # type: ignore
+        def my_handle_wsgi():
+            pass
